recode

+++ myjdbc-example/src/org/dc/jdbc/example/entity/User.java
@@ -3,15 +3,18 @@ package org.dc.jdbc.example.entity;
 import java.util.Date;
 
 public class User {
-	private Integer id;
+	private Long id;
 	private String realName;
 	private Integer sex;
 	private Integer age;
 	private Date createTime;
-	public Integer getId() {
+
+
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getRealName() {
